requirements.txt
streamlit
moviepy<2.0
python-dotenv
Pillow<10.0.0